fix: Overhauls antimacro system
